Restore Redis 6.0 compatibility (do not require 6.2+)
This change is taken from GitLab's modified version to achieve this, see
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/142986,
with a slight modification to remove the logger warning for Redis 6.2,
and instead raise an error on Redis 6.0, which is required.

The motivation for this is mainly to keep supporting the Redis bundled
with Ubuntu 22.04 LTS, as well as Azure Redis service.